@@ -1,7 +1,8 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.0</TargetFramework>
+		<IncludeOpenAPIAnalyzers>true</IncludeOpenAPIAnalyzers>
         <DefaultItemExcludes>$(DefaultItemExcludes);$(SpaRoot)node_modules\**</DefaultItemExcludes>
 
         <!-- Set this to true if you enable server-side prerendering -->
@@ -14,16 +15,26 @@
         <NoWarn>$(NoWarn);1591</NoWarn>
     </PropertyGroup>
 
+    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+      <DocumentationFile>obj\Debug\netcoreapp3.0\AspDotnetVueJs.xml</DocumentationFile>
+      <OutputPath></OutputPath>
+    </PropertyGroup>
+
     <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.App" />
-        <PackageReference Include="Microsoft.AspNetCore.Mvc.Api.Analyzers" Version="2.2.0" />
-        <PackageReference Include="Serilog" Version="2.8.0" />
-        <PackageReference Include="Serilog.Extensions.Logging" Version="2.0.4" />
-        <PackageReference Include="Serilog.AspNetCore" Version="2.1.1" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.0.0" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="3.0.0" />
+        <PackageReference Include="Microsoft.AspNetCore.SpaServices" Version="3.0.0" />
+        <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="3.0.0" />
+        <PackageReference Include="Serilog" Version="2.9.0" />
+        <PackageReference Include="Serilog.Extensions.Logging" Version="3.0.1" />
+        <PackageReference Include="Serilog.AspNetCore" Version="3.1.0" />
         <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
         <PackageReference Include="Serilog.Sinks.File" Version="4.0.0" />
         <PackageReference Include="Serilog.Sinks.RollingFile" Version="3.3.0" />
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc4" />
+        <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="5.0.0-rc4" />
+        <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="5.0.0-rc4" />
+        <PackageReference Include="Westwind.AspnetCore.LiveReload" Version="0.1.5.4" />
     </ItemGroup>
 
     <ItemGroup>
@@ -60,11 +71,9 @@
 
     <Target Name="RunWebpack" AfterTargets="ComputeFilesToPublish">
         <!-- As part of publishing, ensure the JS resources are freshly built in production mode -->
-        <Exec Command="node -e &quot;console.log('NPM Installing dependencies...')&quot;" />
-        <Exec Command="npm install" />
-        <Exec Command="node -e &quot;console.log('Building in $(Configuration) mode')&quot;" />
-        <Exec Condition=" '$(Configuration)' == 'Production' " Command="npm run build:prod" />
-        <Exec Condition=" '$(Configuration)' == 'Test' " Command="npm run build:test" />
+		<Exec Command="node -e &quot;console.log('NPM Installing dependencies...')&quot;" />
+		<Exec Command="npm install" />
+		<Exec Command="npm run build:prod" />
 
         <!-- Include the newly-built files in the publish output -->
         <ItemGroup>
@@ -75,5 +84,7 @@
             </ResolvedFileToPublish>
         </ItemGroup>
     </Target>
+
+    <ProjectExtensions><VisualStudio><UserProperties appsettings_1json__JsonSchema="" /></VisualStudio></ProjectExtensions>
 
 </Project>
